refactor(backend): solve cronjob hapi file order in execution
