+++ kubernetes/cronjob.yaml
@@ -3,7 +3,7 @@ kind: CronJob
 metadata:
   name: weekly
 spec:
-  schedule: "	0 0 * * 0"
+  schedule: "0 0 * * 0"
   jobTemplate:
     spec:
       template:
@@ -25,7 +25,7 @@ kind: CronJob
 metadata:
   name: monthly
 spec:
-  schedule: "	0 0 1 * *"
+  schedule: "0 0 1 * *"
   jobTemplate:
     spec:
       template:
